requirements.txt
Django
djangorestframework
numpy
scikit_learn
sentence_transformers
spacy
supabase
openai
django-cors-headers
nltk
python-dotenv
gunicorn
whitenoise
PyPDF2
huggingface_hub[hf_xet]
hf_transfer
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl